+++ user_payments/user_subscriptions/migrations/0001_initial.py
@@ -53,6 +53,7 @@ class Migration(migrations.Migration):
                     models.CharField(
                         choices=[
                             ("yearly", "yearly"),
+                            ("quarterly", "quarterly"),
                             ("monthly", "monthly"),
                             ("weekly", "weekly"),
                             ("manually", "manually"),